added conditionals to make it easier to not run certain sections

+++ combine_piecemeal/setup_and_run.sh
@@ -1,34 +1,52 @@
 #!/bin/bash
 
 
-echo "drop table range_numbers" | psql cms_query_dev 2> /dev/null
-#cat ddl/range_numbers.ddl | psql cms_query_dev
-cat range_numbers.sql | psql cms_query_dev
-cat test_range_numbers.sql | psql cms_query_dev
+if (TRUE) ;then
+    echo "TEST test_range_numbers.sql"
+    echo "drop table range_numbers" | psql cms_query_dev 2> /dev/null
+    #cat ddl/range_numbers.ddl | psql cms_query_dev
+    cat range_numbers.sql | psql cms_query_dev
+    cat test_range_numbers.sql | psql cms_query_dev
+fi
 
-echo "drop table date_numbers" | psql cms_query_dev 2> /dev/null
-#cat ddl/date_numbers.ddl | psql cms_query_dev
-cat date_numbers.sql | psql cms_query_dev
-cat test_date_numbers.sql | psql cms_query_dev
+if (TRUE) ;then
+    echo "TEST test_date_numbers.sql"
+    echo "drop table date_numbers" | psql cms_query_dev 2> /dev/null
+    #cat ddl/date_numbers.ddl | psql cms_query_dev
+    cat date_numbers.sql | psql cms_query_dev
+    cat test_date_numbers.sql | psql cms_query_dev
+fi
 
-echo "drop table tagged_rows" | psql cms_query_dev 2> /dev/null
-#cat ddl/tagged_rows.ddl | psql cms_query_dev
-cat tagged_rows.sql | psql cms_query_dev
-cat test_tagged_rows.sql | psql cms_query_dev
+if (TRUE) ;then
+    echo "TEST test_tagged_rows.sql"
+    echo "drop table tagged_rows" | psql cms_query_dev 2> /dev/null
+    #cat ddl/tagged_rows.ddl | psql cms_query_dev
+    cat tagged_rows.sql | psql cms_query_dev
+    cat test_tagged_rows.sql | psql cms_query_dev
+fi
 
-echo "drop table end_rows" | psql cms_query_dev 2> /dev/null
-#cat ddl/end_rows.ddl | psql cms_query_dev
-cat end_rows.sql | psql cms_query_dev
-cat test_end_rows.sql | psql cms_query_dev
+if (TRUE) ;then
+    echo "TEST test_end_rows.sql"
+    echo "drop table end_rows" | psql cms_query_dev 2> /dev/null
+    #cat ddl/end_rows.ddl | psql cms_query_dev
+    cat end_rows.sql | psql cms_query_dev
+    cat test_end_rows.sql | psql cms_query_dev
+fi
 
-echo "drop table start_rows" | psql cms_query_dev 2> /dev/null
-#cat ddl/start_rows.ddl | psql cms_query_dev
-cat start_rows.sql | psql cms_query_dev
-cat test_start_rows.sql | psql cms_query_dev
+if (TRUE) ;then
+    echo "TEST test_start_rows.sql"
+    echo "drop table start_rows" | psql cms_query_dev 2> /dev/null
+    #cat ddl/start_rows.ddl | psql cms_query_dev
+    cat start_rows.sql | psql cms_query_dev
+    cat test_start_rows.sql | psql cms_query_dev
+fi
 
-echo "drop table ranges" | psql cms_query_dev
-#cat ddl/ranges.ddl | psql cms_query_dev
-cat ranges.sql | psql cms_query_dev
-cat test_ranges.sql | psql cms_query_dev
+if (TRUE) ;then
+    echo "TEST test_ranges.sql"
+    echo "drop table ranges" | psql cms_query_dev
+    #cat ddl/ranges.ddl | psql cms_query_dev
+    cat ranges.sql | psql cms_query_dev
+    cat test_ranges.sql | psql cms_query_dev
+fi
